+++ lib/pay/stripe/billable.rb
@@ -22,11 +22,25 @@ module Pay
         @pay_customer = pay_customer
       end
 
+      def customer_metadata
+        owner = pay_customer.owner
+        case owner.class.pay_customer_metadata
+        when Symbol
+          owner.send(owner.class.pay_customer_metadata, pay_customer)
+        when Proc
+          owner.class.pay_customer_metadata.call(pay_customer)
+        end
+      end
+
       def customer
         stripe_customer = if processor_id?
           ::Stripe::Customer.retrieve({id: processor_id}, stripe_options)
         else
-          sc = ::Stripe::Customer.create({email: email, name: customer_name}, stripe_options)
+          sc = ::Stripe::Customer.create({
+            email: email,
+            name: customer_name,
+            metadata: customer_metadata
+          }, stripe_options)
           pay_customer.update!(processor_id: sc.id, stripe_account: stripe_account)
           sc
         end
@@ -47,7 +61,15 @@ module Pay
       # Syncs name and email to Stripe::Customer
       def update_customer!
         return unless processor_id?
-        ::Stripe::Customer.update(processor_id, {name: customer_name, email: email}, stripe_options)
+        ::Stripe::Customer.update(
+          processor_id,
+          {
+            email: email,
+            name: customer_name,
+            metadata: customer_metadata
+          },
+          stripe_options
+        )
       end
 
       def charge(amount, options = {})